+++ parser/config.go
@@ -144,6 +144,7 @@ var SupportedFunctions = map[string]int{
 	"primorial": 1,
 	"lambertw":  1,
 	"zeta":      1,
+	"primecount": 1,
 	"digitcount": 1, // Collatz stopping time
 	"prod":      -1,
 	"min":       -1, // variadic